# lib/api-zod/src/generated/types/researchSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchSessionStatus } from "./researchSessionStatus";
import type { ResearchStep } from "./researchStep";

export interface ResearchSession {
  id: string;
  topic: string;
  status: ResearchSessionStatus;
  steps: ResearchStep[];
  report?: string;
  createdAt: string;
  completedAt?: string;
}