# lib/api-zod/src/generated/types/researchStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchSource } from "./researchSource";
import type { ResearchStepType } from "./researchStepType";

export interface ResearchStep {
  type: ResearchStepType;
  subQuestion?: string;
  content: string;
  sources?: ResearchSource[];
  timestamp: string;
}